Move aws.S3.role to aws.role
The output for AWS supports S3 and Cloudwatch, but the role is only defined for
S3. So the role should be a level higher.

The implementation in this commit, will support both role levels, as I didn't
want to break current config.

This commit will use `fluentd.aws.role`, and if that isn't set then it will use
`fluentd.aws.S3.role`. As a fall back it will set the role as
`NO_ROLE_SET_CHECK_VALUES`